Switch affiliate.js to be in ldsg.js
Fixes some small things that Bobby told me to. Should work
